requirements.txt
python-docx===1.1.2
python-pptx===1.0.2
pandas===2.2.2
reportlab===4.2.2
openpyxl===3.1.5
pyside6===6.7.2
pillow===10.4.0
markdown2===2.5.0